fixed multiscale preprocess_batch

+++ ultralytics/models/yolo/detect/train.py
@@ -60,9 +60,9 @@ class DetectionTrainer(BaseTrainer):
         if self.args.multi_scale:
             imgs = batch["img"]
             sz = (
-                random.randrange(self.args.imgsz * 0.5, self.args.imgsz * 1.5 + self.stride)
+                random.randrange(int(self.args.imgsz * 0.5), int(self.args.imgsz * 1.5 + self.stride)
                 // self.stride
-                * self.stride
+                * self.stride)
             )  # size
             sf = sz / max(imgs.shape[2:])  # scale factor
             if sf != 1:
